Add files via upload
This update includes a game timer and heart system for the player that loses one life when the guard(frog) collides with the player. One power up (extra heart) is included as well that adds one life to the player. Teleporting between scenes has also been added. When the player collides with the door (eagle), they will go to level 2. 
Date - 3/2/19